Refactored server api
